cgcnn

+++ nn_utils/cgcnn.py
@@ -149,7 +149,7 @@ class CrystalGraphConvNet(nn.Module):
 
         """
 
-        atom_fea = graph.ndata['atomic_number'].reshape(-1,1)
+        atom_fea = graph.ndata['atomic_number'].reshape(-1,1).to(torch.float32)
         nbr_fea = graph.ndata['gaussian_dist']
         nbr_fea_idx = graph.ndata['nbr_fea_idx']
         batch = graph.batch_num_nodes()
